Crate namada_gas

Gas accounting module to track the gas usage in a block for transactions and validity predicates triggered by transactions.

Modules

event

Gas related events.

storage

Gas storage keys

Structs

Gas

Representation of tracking gas in sub-units.

TxGasMeter

Gas metering in a transaction

VpGasMeter

Gas metering in a validity predicate

WholeGas

Gas represented in whole units. Used for fee payment and to display information to the user.

Enums

Error

GasParseError

Constants

IBC_ACTION_EXECUTE_GAS

The cost to execute an Ibc action

IBC_ACTION_VALIDATE_GAS

The cost to validate an Ibc action

MASP_CONVERT_CHECK_GAS

The cost to process a masp convert note in the bundle

MASP_FINAL_CHECK_GAS

The cost to run the final masp check in the bundle

MASP_FIXED_CONVERT_GAS

The fixed cost of convert note verification

MASP_FIXED_OUTPUT_GAS

The fixed cost of output note verification

MASP_FIXED_SPEND_GAS

The fixed cost of spend note verification

MASP_OUTPUT_CHECK_GAS

The cost to process a masp output note in the bundle

MASP_SPEND_CHECK_GAS

The cost to process a masp spend note in the bundle

MASP_VARIABLE_CONVERT_GAS

The variable cost of convert note verification

MASP_VARIABLE_OUTPUT_GAS

The variable cost of output note verification

MASP_VARIABLE_SPEND_GAS

The variable cost of spend note verification

MASP_VERIFY_SIG_GAS

The cost of masp sig verification

MEMORY_ACCESS_GAS_PER_BYTE

The cost of accessing data from memory (both read and write mode), per byte

STORAGE_ACCESS_GAS_PER_BYTE

The cost of accessing data from storage, per byte

STORAGE_DELETE_GAS_PER_BYTE

The cost of removing data from storage, per byte

STORAGE_WRITE_GAS_PER_BYTE

The cost of writing data to storage, per byte

VERIFY_TX_SIG_GAS

The cost of verifying a single signature of a transaction

WASM_MEMORY_PAGE_GAS

The cost for requesting one more page in wasm (64KiB)

Traits

GasMetering

Trait to share gas operations for transactions and validity predicates

Type Aliases

Result

Gas module result for functions that may fail