namada_gas
Module storage

Gas storage keys

Structs

Key

A storage key is made of storage key segments DbKeySeg, separated by KEY_SEGMENT_SEPARATOR.

Functions

gas_prefix

Get the full transaction hash prefix under the last subkey

pred_prefix

Get the transaction hash prefix under the all subkey