namada_gas
Trait GasMetering

pub trait GasMetering {
    // Required methods
    fn consume(&mut self, gas: Gas) -> Result<()>;
    fn get_tx_consumed_gas(&self) -> Gas;
    fn get_gas_limit(&self) -> Gas;

    // Provided methods
    fn add_compiling_gas(&mut self, bytes_len: u64) -> Result<()> { ... }
    fn add_wasm_load_from_storage_gas(&mut self, bytes_len: u64) -> Result<()> { ... }
    fn add_wasm_validation_gas(&mut self, bytes_len: u64) -> Result<()> { ... }
    fn check_vps_limit(&self, vps_gas: Gas) -> Result<()> { ... }
}

Trait to share gas operations for transactions and validity predicates

Required Methods

fn consume(&mut self, gas: Gas) -> Result<()>

Add gas cost. It will return error when the consumed gas exceeds the provided transaction gas limit, but the state will still be updated

fn get_tx_consumed_gas(&self) -> Gas

Get the gas consumed by the tx

fn get_gas_limit(&self) -> Gas

Get the gas limit

Provided Methods

fn add_compiling_gas(&mut self, bytes_len: u64) -> Result<()>

Add the compiling cost proportionate to the code length

fn add_wasm_load_from_storage_gas(&mut self, bytes_len: u64) -> Result<()>

Add the gas for loading the wasm code from storage

fn add_wasm_validation_gas(&mut self, bytes_len: u64) -> Result<()>

Add the gas for validating untrusted wasm code

fn check_vps_limit(&self, vps_gas: Gas) -> Result<()>

Check if the vps went out of gas. Starts with the gas consumed by the transaction.